chore(tauri) rename tauri to tauri-api and tauri-bundle to tauri
