+++ tauri-api/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod command;
+pub mod dir;
+pub mod file;
+pub mod rpc;
+
+pub mod version;
