+++ tauri-api/src/version/error.rs
