--- tauri-build/Cargo.toml
@@ -1,48 +1,0 @@
-[package]
-name = "tauri-build"
-version = "0.2.0"
-authors = ["Lucas Fernandes Gonçalves Nogueira <lucas@quasar.dev>", "Daniel Thompson-Yvetot <denjell@sfosc.org>", "Tensor Programming <tensordeveloper@gmail.com>"]
-license = "MIT"
-homepage = "https://tauri-apps.org"
-repository = "https://github.com/tauri-apps/tauri"
-description = "Make tiny, secure apps for all desktop platforms with Tauri"
-edition = "2018"
-exclude = ["test/fixture/**"]
-
-[dependencies]
-serde_json = "1.0.44"
-serde = "1.0"
-serde_derive = "1.0"
-rand = "0.7"
-web-view = "0.5.4"
-tauri_includedir = "0.5.0"
-phf = "0.8.0"
-base64 = "0.11.0"
-webbrowser = "0.5.2"
-lazy_static = "1.4.0"
-tiny_http = "0.6"
-threadpool = "1.7"
-uuid = { version = "0.8.1", features = ["v4"] }
-tauri = { version = "0.2",  path = "../tauri" }
-
-[build-dependencies]
-tauri_includedir_codegen = "0.5.1"
-serde_json = "1.0.44"
-serde = "1.0"
-serde_derive = "1.0"
-
-[features]
-edge = ["web-view/edge"]
-dev-server = []
-embedded-server = []
-no-server = []
-all-api = []
-readTextFile = []
-readBinaryFile = []
-writeFile = []
-listFiles = []
-listDirs = []
-setTitle = []
-execute = []
-open = []
-emit = []
