+++ tauri/Cargo.toml
@@ -10,13 +10,39 @@ edition = "2018"
 exclude = ["test/fixture/**"]
 
 [dependencies]
-serde = { version = "1.0", features = ["derive"] } 
-dirs = "2.0.2"
-ignore = "0.4.10"
-zip = "0.5.3"
-tempdir = "0.3"
-semver = "0.9"
-tempfile = "3"
-either = "1.5.3"
-tar = "0.4"
-flate2 = "1"
+serde_json = "1.0.44"
+serde = "1.0"
+serde_derive = "1.0"
+rand = "0.7"
+web-view = "0.5.4"
+tauri_includedir = "0.5.0"
+phf = "0.8.0"
+base64 = "0.11.0"
+webbrowser = "0.5.2"
+lazy_static = "1.4.0"
+tiny_http = "0.6"
+threadpool = "1.7"
+uuid = { version = "0.8.1", features = ["v4"] }
+tauri-api = { version = "0.2",  path = "../tauri-api" }
+
+[build-dependencies]
+tauri_includedir_codegen = "0.5.1"
+serde_json = "1.0.44"
+serde = "1.0"
+serde_derive = "1.0"
+
+[features]
+edge = ["web-view/edge"]
+dev-server = []
+embedded-server = []
+no-server = []
+all-api = []
+readTextFile = []
+readBinaryFile = []
+writeFile = []
+listFiles = []
+listDirs = []
+setTitle = []
+execute = []
+open = []
+emit = []
