+++ tauri/src/endpoints/cmd.rs
