+++ tauri/src/salt.rs
